refactor(client): extract classification tests to sibling file
